+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Changelog
 
+## [Upcoming]
+* `rasterise` function now can work with whole ggplot objects
+
 ## [1.0.0] - 2021-11-15
 * `ggrastr` has gone through no major revisions in over a year. In order to avoid any confusion, this should be released with a major version.
 
 ## [0.2.3] - 2021-02-27
 * Function `rasterise()` changed to work with multiple layers
 * Function `rasterise()` now works with `geom_sf()`, i.e. should find any layers in a valid input list, and rasterize this. 
-* Parameter `scale` added to rasterise()`. This scales the 'height' and 'weight' of the raster objects
+* Parameter `scale` added to `rasterise()`. This scales the 'height' and 'weight' of the raster objects
 * Updates to roxygen2, added types
 
 ## [0.2.2] - 2021-02-10